refactor(hooks): update hooks
