+++ src/hooks/useCharacters.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { getCharacters } from '@/core/api/getCharacters.ts';
-import type { TApiResponse } from '@/shema/characterShema.ts';
+import type { TApiResponse } from '@/scheme/characterScheme.ts';
 
 export const useCharacters = (page: number, query: string) => {
   return useQuery<TApiResponse, Error>({
